add-on to htms-prediction.js
I have included the add-on I mentioned in the HT forum. I called it Mimimi add-on.
It is very simple, as you can see. The comments in the code explain how it works.
Please, feel free to include it in the official version of Foxtrick, as well as, to modify the way you want to show the content.