Fix lastlineup redirection
The order of matchs in the Matches page was reversed recently. Thus, we need to use the first lineup, not the last.

Interestingly, the first hit of querySelector is not the first lineup in the list, but the one in the featured last match / next match at the very top of the page.