Fix redir_to_coach (open specialist page)
The recent coach reform changed the Hattrick-generated link from the player page to the specialist page.
The player page barely contain any information about the coach status (salary + link to specialist page), so there is no point in opening it.

Trying to redirect to a null URL is a no-op operation, so in case of error, we would only open the Players page.