[math.js] Fix Predict.stamina for players at exactly 8.00 stamina
Players at exactly stamina 8.00 (that is, 33 y.o. players) are well-known to be displayed as having 7.98 stamina in MatchLineupTweaks, despite having a visible 8 stamina in Hattrick.
Tweaking the formula impacts players at 8.00 stamina and those slightly above. For instance, energyAt90 === 0.88699 are upgraded from a computed stamina of 7.9989 to 8.0111, which may or may not be correct, but at least is in 8+ territory.
It should also be noted that the formula change does not try to keep the function continuous at its breakpoint, as `Foxtrick.Predict.stamina(0.88499)` now returns 7.978677159, while it most likely should be way closer to 8.00.